refactor: migrate to BigDecimal and fix product integration

+++ src/main/java/com/ecommerce/cart_service/dto/CartItemResponse.java
@@ -1,10 +1,12 @@
 package com.ecommerce.cart_service.dto;
 
+import java.math.BigDecimal;
+
 public record CartItemResponse(
-        String productId,
-        int quantity,
-        double price,
-        String imageUrl,
-        String productName,
-        Boolean available) {
+                String productId,
+                int quantity,
+                BigDecimal price,
+                String imageUrl,
+                String productName,
+                Boolean available) {
 }
